Refactor MDS client code into modular files
