+++ examples/server/server/static/mds.js
@@ -1,238 +1,30 @@
-const MDS_HTML_PATH = 'mds.html';
-const MDS_JWS_PATH = 'fido-mds3.jws';
-const COLUMN_COUNT = 11;
-const MISSING_METADATA_MESSAGE = 'Metadata has not been downloaded yet. Use the Download Metadata button.';
-
-const UPDATE_BUTTON_STATES = {
-    update: { label: 'Update Metadata', busyLabel: 'Updating…' },
-    download: { label: 'Download Metadata', busyLabel: 'Downloading…' },
-};
-
-const CERTIFICATION_OPTIONS = [
-    'FIDO_CERTIFIED',
-    'FIDO_CERTIFIED_L1',
-    'FIDO_CERTIFIED_L2',
-    'NOT_FIDO_CERTIFIED',
-    'REVOKED',
-];
-
-const FILTER_CONFIG = [
-    { key: 'name', inputId: 'mds-filter-name' },
-    { key: 'protocol', inputId: 'mds-filter-protocol', optionsKey: 'protocol' },
-    {
-        key: 'certification',
-        inputId: 'mds-filter-certification',
-        optionsKey: 'certification',
-        staticOptions: CERTIFICATION_OPTIONS,
-    },
-    { key: 'id', inputId: 'mds-filter-id' },
-    {
-        key: 'userVerification',
-        inputId: 'mds-filter-user-verification',
-        optionsKey: 'userVerification',
-        expandDropdown: true,
-    },
-    { key: 'attachment', inputId: 'mds-filter-attachment', optionsKey: 'attachment' },
-    { key: 'transports', inputId: 'mds-filter-transports', optionsKey: 'transports' },
-    { key: 'keyProtection', inputId: 'mds-filter-key-protection', optionsKey: 'keyProtection' },
-    {
-        key: 'algorithms',
-        inputId: 'mds-filter-algorithms',
-        optionsKey: 'algorithms',
-        expandDropdown: true,
-    },
-];
-
-const FILTER_LOOKUP = FILTER_CONFIG.reduce((map, config) => {
-    map[config.key] = config;
-    return map;
-}, {});
-
-let activeDropdown = null;
-
-class FilterDropdown {
-    constructor(input, onSelect, config = {}) {
-        this.input = input;
-        this.onSelect = onSelect;
-        this.options = [];
-        this.filtered = [];
-        this.activeIndex = -1;
-        this.list = null;
-        this.container = null;
-        this.expandToContent = Boolean(config.expandDropdown);
-
-        const parent = input.parentElement;
-        if (parent) {
-            parent.classList.add('mds-filter-cell');
-        }
-
-        this.container = document.createElement('div');
-        this.container.className = 'mds-filter-dropdown';
-        this.container.hidden = true;
-        if (this.expandToContent) {
-            this.container.classList.add('mds-filter-dropdown--expanded');
-        }
-
-        this.list = document.createElement('ul');
-        this.list.className = 'mds-filter-dropdown__list';
-        this.container.appendChild(this.list);
-
-        if (parent) {
-            parent.appendChild(this.container);
-        } else {
-            input.insertAdjacentElement('afterend', this.container);
-        }
-
-        this.handleDocumentClick = this.handleDocumentClick.bind(this);
-
-        input.addEventListener('focus', () => this.open());
-        input.addEventListener('click', () => this.open());
-        input.addEventListener('keydown', event => this.handleKeyDown(event));
-        input.addEventListener('input', () => this.filter(this.input.value));
-
-        this.container.addEventListener('mousedown', event => {
-            event.preventDefault();
-        });
-
-        this.container.addEventListener('click', event => {
-            const optionEl = event.target.closest('.mds-filter-dropdown__option');
-            if (optionEl) {
-                const value = optionEl.getAttribute('data-value') || optionEl.textContent || '';
-                this.select(value);
-            }
-        });
-
-        document.addEventListener('mousedown', this.handleDocumentClick);
-    }
-
-    setOptions(options) {
-        const unique = Array.from(new Set(options.filter(Boolean)));
-        unique.sort((a, b) => a.localeCompare(b, undefined, { sensitivity: 'base' }));
-        this.options = unique;
-        this.filter(this.input.value);
-    }
-
-    open() {
-        if (!this.options.length) {
-            return;
-        }
-        if (activeDropdown && activeDropdown !== this) {
-            activeDropdown.close();
-        }
-        activeDropdown = this;
-        this.container.hidden = false;
-        this.container.classList.add('is-open');
-        this.activeIndex = -1;
-        this.filter(this.input.value);
-    }
-
-    close() {
-        if (activeDropdown === this) {
-            activeDropdown = null;
-        }
-        this.container.hidden = true;
-        this.container.classList.remove('is-open');
-        this.activeIndex = -1;
-    }
-
-    filter(query) {
-        const value = (query || '').trim().toLowerCase();
-        if (!value) {
-            this.filtered = [...this.options];
-        } else {
-            this.filtered = this.options.filter(option => option.toLowerCase().includes(value));
-        }
-        this.render();
-    }
-
-    render() {
-        if (!this.list) {
-            return;
-        }
-        this.list.innerHTML = '';
-
-        const items = this.filtered.length ? this.filtered : [];
-        if (!items.length) {
-            if (!this.options.length) {
-                return;
-            }
-            const empty = document.createElement('li');
-            empty.className = 'mds-filter-dropdown__empty';
-            empty.textContent = 'No matches';
-            this.list.appendChild(empty);
-            return;
-        }
-
-        items.forEach((option, index) => {
-            const item = document.createElement('li');
-            item.className = 'mds-filter-dropdown__option';
-            if (index === this.activeIndex) {
-                item.classList.add('is-active');
-            }
-            item.textContent = option;
-            item.setAttribute('data-value', option);
-            this.list.appendChild(item);
-        });
-    }
-
-    handleKeyDown(event) {
-        if (!this.options.length) {
-            return;
-        }
-        if (event.key === 'Escape') {
-            this.close();
-            return;
-        }
-        if (event.key === 'ArrowDown' || event.key === 'ArrowUp') {
-            if (!this.container.classList.contains('is-open')) {
-                this.open();
-            }
-            const delta = event.key === 'ArrowDown' ? 1 : -1;
-            this.move(delta);
-            event.preventDefault();
-            return;
-        }
-        if (event.key === 'Enter' && this.container.classList.contains('is-open') && this.activeIndex >= 0) {
-            const items = this.filtered.length ? this.filtered : this.options;
-            const value = items[this.activeIndex];
-            if (value) {
-                this.select(value);
-                event.preventDefault();
-            }
-        }
-    }
-
-    move(delta) {
-        const items = this.filtered.length ? this.filtered : this.options;
-        if (!items.length) {
-            this.activeIndex = -1;
-            this.render();
-            return;
-        }
-        this.activeIndex = (this.activeIndex + delta + items.length) % items.length;
-        this.render();
-        const activeEl = this.list?.querySelector('.mds-filter-dropdown__option.is-active');
-        if (activeEl) {
-            activeEl.scrollIntoView({ block: 'nearest' });
-        }
-    }
-
-    select(value) {
-        this.input.value = value;
-        this.onSelect(value);
-        this.close();
-    }
-
-    handleDocumentClick(event) {
-        if (!this.container.contains(event.target) && event.target !== this.input) {
-            this.close();
-        }
-    }
-}
-
-function createFilterDropdown(input, onSelect, config = {}) {
-    return new FilterDropdown(input, onSelect, config);
-}
+import {
+    MDS_HTML_PATH,
+    MDS_JWS_PATH,
+    COLUMN_COUNT,
+    MISSING_METADATA_MESSAGE,
+    UPDATE_BUTTON_STATES,
+    FILTER_CONFIG,
+    FILTER_LOOKUP,
+} from './mds-constants.js';
+import { createFilterDropdown } from './mds-dropdown.js';
+import {
+    collectOptionSets,
+    transformEntry,
+    parseIsoDate,
+    formatDate,
+    decodeBase64Url,
+    formatEnum,
+    normaliseEnumKey,
+    normaliseAaguid,
+    formatListValues,
+    formatDetailValue,
+    formatAuthenticatorAlgorithms,
+    formatGuidCandidate,
+    formatUpv,
+    extractList,
+    renderCertificateSummary,
+} from './mds-utils.js';
 
 let mdsState = null;
 let mdsData = [];
@@ -1720,265 +1512,6 @@ function formatCertificateOutput(details) {
     return JSON.stringify(details, null, 2);
 }
 
-function formatCertificateDateDisplay(value) {
-    if (!value) {
-        return '';
-    }
-    const date = new Date(value);
-    if (!Number.isNaN(date.getTime())) {
-        return date.toUTCString();
-    }
-    return typeof value === 'string' ? value : '';
-}
-
-function createSummaryItem(label, value, options = {}) {
-    if (!label) {
-        return null;
-    }
-
-    const resolved = Array.isArray(value) ? value.filter(Boolean) : value;
-    const isArray = Array.isArray(resolved);
-    const scalar = !isArray ? resolved : null;
-    const text = typeof scalar === 'string' ? scalar.trim() : scalar;
-
-    if ((!isArray && (text === undefined || text === null || text === '')) || (isArray && !resolved.length)) {
-        return null;
-    }
-
-    const item = document.createElement('li');
-    item.className = 'mds-certificate-summary__item';
-
-    const labelEl = document.createElement('div');
-    labelEl.className = 'mds-certificate-summary__label';
-    labelEl.textContent = label;
-    item.appendChild(labelEl);
-
-    const valueEl = document.createElement('div');
-    valueEl.className = 'mds-certificate-summary__value';
-
-    if (options.code) {
-        const codeEl = document.createElement('code');
-        codeEl.className = 'mds-certificate-summary__code';
-        codeEl.textContent = String(value);
-        valueEl.appendChild(codeEl);
-    } else if (isArray) {
-        resolved.forEach(entry => {
-            const line = document.createElement('div');
-            line.textContent = String(entry);
-            valueEl.appendChild(line);
-        });
-    } else {
-        valueEl.textContent = String(text);
-    }
-
-    item.appendChild(valueEl);
-    return item;
-}
-
-function determinePublicKeyAlgorithm(info) {
-    if (!info || typeof info !== 'object') {
-        return '';
-    }
-    const algorithm = info.algorithm;
-    if (algorithm) {
-        if (typeof algorithm === 'string') {
-            const algorithmName = algorithm.trim();
-            if (algorithmName) {
-                return algorithmName;
-            }
-        }
-        if (typeof algorithm === 'object') {
-            const name = typeof algorithm.name === 'string' ? algorithm.name.trim() : '';
-            if (name) {
-                return name;
-            }
-        }
-    }
-    const type = typeof info.type === 'string' ? info.type.trim() : '';
-    return type;
-}
-
-function formatSignatureHashName(hash) {
-    if (typeof hash !== 'string') {
-        return '';
-    }
-    const trimmed = hash.trim();
-    if (!trimmed) {
-        return '';
-    }
-    const simpleShaMatch = /^sha(\d{3})$/i.exec(trimmed);
-    if (simpleShaMatch) {
-        return `SHA-${simpleShaMatch[1]}`;
-    }
-    return trimmed.toUpperCase();
-}
-
-function renderCertificatePublicKey(info) {
-    if (!info || typeof info !== 'object') {
-        return null;
-    }
-
-    const section = document.createElement('div');
-    section.className = 'mds-certificate-summary__section';
-
-    const title = document.createElement('div');
-    title.className = 'mds-certificate-summary__label';
-    title.textContent = 'Public Key';
-    section.appendChild(title);
-
-    const list = document.createElement('ul');
-    list.className = 'mds-certificate-summary__list';
-
-    const algorithmItem = createSummaryItem('Algorithm', determinePublicKeyAlgorithm(info));
-    if (algorithmItem) {
-        list.appendChild(algorithmItem);
-    }
-
-    const algorithmDetails = info.algorithm && typeof info.algorithm === 'object' ? info.algorithm : null;
-    const curveValue = info.curve || (algorithmDetails && algorithmDetails.namedCurve);
-    if (curveValue) {
-        const curveItem = createSummaryItem('Named Curve', curveValue);
-        if (curveItem) {
-            list.appendChild(curveItem);
-        }
-    }
-
-    const modulusLength = algorithmDetails && algorithmDetails.modulusLength;
-    const keySize = modulusLength || info.keySize;
-    if (keySize) {
-        const sizeItem = createSummaryItem('Key Size', `${keySize} bit`);
-        if (sizeItem) {
-            list.appendChild(sizeItem);
-        }
-    }
-
-    if (info.publicExponent !== undefined && info.publicExponent !== null) {
-        const exponentItem = createSummaryItem('Public Exponent', String(info.publicExponent));
-        if (exponentItem) {
-            list.appendChild(exponentItem);
-        }
-    }
-
-    if (info.modulusHex) {
-        const modulusItem = createSummaryItem('Modulus', info.modulusHex, { code: true });
-        if (modulusItem) {
-            list.appendChild(modulusItem);
-        }
-    }
-
-    if (info.uncompressedPoint) {
-        const pointItem = createSummaryItem('Uncompressed Point', info.uncompressedPoint, { code: true });
-        if (pointItem) {
-            list.appendChild(pointItem);
-        }
-    }
-
-    if (info.subjectPublicKeyInfoBase64) {
-        const valueItem = createSummaryItem('Value', info.subjectPublicKeyInfoBase64, { code: true });
-        if (valueItem) {
-            list.appendChild(valueItem);
-        }
-    }
-
-    if (!list.childElementCount) {
-        return null;
-    }
-
-    section.appendChild(list);
-    return section;
-}
-
-function renderCertificateSignature(signature) {
-    if (!signature || typeof signature !== 'object') {
-        return null;
-    }
-
-    const section = document.createElement('div');
-    section.className = 'mds-certificate-summary__section';
-
-    const title = document.createElement('div');
-    title.className = 'mds-certificate-summary__label';
-    title.textContent = 'Signature';
-    section.appendChild(title);
-
-    const list = document.createElement('ul');
-    list.className = 'mds-certificate-summary__list';
-
-    if (signature.algorithm) {
-        const algorithmItem = createSummaryItem('Algorithm', signature.algorithm);
-        if (algorithmItem) {
-            list.appendChild(algorithmItem);
-        }
-    }
-
-    if (signature.hash) {
-        const hashName = typeof signature.hash === 'object' && signature.hash !== null
-            ? signature.hash.name
-            : signature.hash;
-        const hashValue = typeof hashName === 'string' ? formatSignatureHashName(hashName) : hashName;
-        const hashItem = createSummaryItem('Hash', hashValue);
-        if (hashItem) {
-            list.appendChild(hashItem);
-        }
-    }
-
-    if (signature.hex) {
-        const valueItem = createSummaryItem('Value', signature.hex, { code: true });
-        if (valueItem) {
-            list.appendChild(valueItem);
-        }
-    }
-
-    if (!list.childElementCount) {
-        return null;
-    }
-
-    section.appendChild(list);
-    return section;
-}
-
-function renderCertificateSummary(details) {
-    if (!details || typeof details !== 'object') {
-        return null;
-    }
-
-    const fragment = document.createDocumentFragment();
-
-    const infoList = document.createElement('ul');
-    infoList.className = 'mds-certificate-summary__list';
-
-    const validity = details.validity || {};
-    const serialNumber = details.serialNumber || {};
-
-    [
-        createSummaryItem('Subject', details.subject),
-        createSummaryItem('Issuer', details.issuer),
-        createSummaryItem('Not Before', formatCertificateDateDisplay(validity.notBefore)),
-        createSummaryItem('Not After', formatCertificateDateDisplay(validity.notAfter)),
-        createSummaryItem('Serial Number', serialNumber.decimal || serialNumber.hex),
-        serialNumber.hex ? createSummaryItem('Serial Number (Hex)', serialNumber.hex) : null,
-    ].forEach(item => {
-        if (item) {
-            infoList.appendChild(item);
-        }
-    });
-
-    if (infoList.childElementCount) {
-        fragment.appendChild(infoList);
-    }
-
-    const publicKeySection = renderCertificatePublicKey(details.publicKeyInfo);
-    if (publicKeySection) {
-        fragment.appendChild(publicKeySection);
-    }
-
-    const signatureSection = renderCertificateSignature(details.signature);
-    if (signatureSection) {
-        fragment.appendChild(signatureSection);
-    }
-
-    return fragment.childElementCount ? fragment : null;
-}
 
 function setCertificateSummaryContent(content) {
     if (!mdsState?.certificateSummary) {
@@ -2572,34 +2105,6 @@ async function refreshMetadata() {
     }
 }
 
-function collectOptionSets(data) {
-    const sets = {
-        protocol: new Set(),
-        certification: new Set(CERTIFICATION_OPTIONS.map(option => formatEnum(option))),
-        userVerification: new Set(),
-        attachment: new Set(),
-        transports: new Set(),
-        keyProtection: new Set(),
-        algorithms: new Set(),
-    };
-
-    data.forEach(entry => {
-        if (entry.protocol) {
-            sets.protocol.add(entry.protocol);
-        }
-        if (entry.certificationStatus) {
-            sets.certification.add(formatEnum(entry.certificationStatus));
-        }
-        entry.userVerificationList.forEach(value => sets.userVerification.add(value));
-        entry.attachmentList.forEach(value => sets.attachment.add(value));
-        entry.transportsList.forEach(value => sets.transports.add(value));
-        entry.keyProtectionList.forEach(value => sets.keyProtection.add(value));
-        entry.algorithmsList.forEach(value => sets.algorithms.add(value));
-    });
-
-    return sets;
-}
-
 function updateOptionLists(optionSets) {
     if (!mdsState) {
         return;
@@ -2623,438 +2128,6 @@ function updateOptionLists(optionSets) {
     });
 }
 
-function transformEntry(entry, index = 0) {
-    const metadata = entry?.metadataStatement ?? {};
-    const name = resolveName(metadata, entry);
-    const protocol = formatProtocol(metadata.protocolFamily || metadata.protocolType);
-    const { display: certification, status: certificationStatus } = formatCertification(entry?.statusReports || []);
-    const identifier = resolveIdentifier(entry, metadata);
-    const aaguid = resolveAaguid(entry, metadata) || '';
-    const userVerificationList = extractUserVerification(metadata.userVerificationDetails);
-    const attachmentList = extractList(metadata.attachmentHint).map(formatEnum);
-    const transportsList = extractTransports(metadata);
-    const keyProtectionList = extractList(metadata.keyProtection).map(formatEnum);
-    const algorithmsList = extractList(metadata.authenticationAlgorithms).map(formatEnum);
-    const icon = normaliseIcon(metadata.icon, metadata.iconType);
-    const attestationCertificates = extractList(metadata.attestationRootCertificates);
-    const attestationKeyIdentifiers = extractAttestationKeyIdentifiers(metadata, entry);
-
-    const latestStatusDate = latestEffectiveDate(entry?.statusReports || []);
-    const rawDate = entry?.timeOfLastStatusChange || latestStatusDate;
-    const dateUpdated = rawDate ? formatDate(rawDate) : '';
-
-    return {
-        index,
-        name,
-        protocol,
-        certification,
-        certificationStatus,
-        id: identifier,
-        aaguid,
-        icon,
-        userVerification: userVerificationList.join(', '),
-        userVerificationList,
-        attachment: attachmentList.join(', '),
-        attachmentList,
-        transports: transportsList.join(', '),
-        transportsList,
-        keyProtection: keyProtectionList.join(', '),
-        keyProtectionList,
-        algorithms: algorithmsList.join(', '),
-        algorithmsList,
-        dateUpdated,
-        dateTooltip: rawDate || undefined,
-        metadataStatement: metadata,
-        rawEntry: entry || null,
-        statusReports: Array.isArray(entry?.statusReports) ? entry.statusReports : [],
-        attestationCertificates,
-        attestationKeyIdentifiers,
-    };
-}
-
-function extractAttestationKeyIdentifiers(metadata, entry) {
-    const map = new Map();
-    const addValue = value => {
-        if (value === undefined || value === null) {
-            return;
-        }
-        const text = String(value).trim();
-        if (!text) {
-            return;
-        }
-        const key = text.toLowerCase();
-        if (!map.has(key)) {
-            map.set(key, text);
-        }
-    };
-
-    extractList(metadata?.attestationCertificateKeyIdentifiers).forEach(addValue);
-    extractList(entry?.attestationCertificateKeyIdentifiers).forEach(addValue);
-
-    return Array.from(map.values());
-}
-
-function normaliseIcon(icon, iconType) {
-    if (!icon) {
-        return '';
-    }
-    const value = String(icon).trim();
-    if (!value) {
-        return '';
-    }
-    if (/^data:/i.test(value)) {
-        return value;
-    }
-    if (/^https?:\/\//i.test(value)) {
-        return value;
-    }
-    const type = typeof iconType === 'string' && iconType.trim() ? iconType.trim() : 'image/png';
-    return `data:${type};base64,${value}`;
-}
-
-function resolveName(metadata, entry) {
-    const description = metadata.description;
-    if (typeof description === 'string' && description.trim()) {
-        return description.trim();
-    }
-    if (description && typeof description === 'object') {
-        const values = Object.values(description).filter(Boolean);
-        if (values.length) {
-            return String(values[0]).trim();
-        }
-    }
-    const altDescriptions = metadata.alternativeDescriptions;
-    if (altDescriptions) {
-        const altValues = typeof altDescriptions === 'object' ? Object.values(altDescriptions) : [];
-        const candidate = altValues.find(value => typeof value === 'string' && value.trim());
-        if (candidate) {
-            return candidate.trim();
-        }
-    }
-    const statusDescriptor = entry?.statusReports?.find(report => report.certificationDescriptor)?.certificationDescriptor;
-    if (statusDescriptor) {
-        return statusDescriptor;
-    }
-    return 'Unknown Authenticator';
-}
-
-function resolveIdentifier(entry, metadata) {
-    if (entry?.aaguid) {
-        return entry.aaguid;
-    }
-    if (metadata?.aaguid) {
-        return metadata.aaguid;
-    }
-    if (metadata?.aaid) {
-        return metadata.aaid;
-    }
-    const attestKeyIds = extractList(metadata?.attestationCertificateKeyIdentifiers);
-    if (attestKeyIds.length) {
-        return attestKeyIds[0];
-    }
-    return '—';
-}
-
-function resolveAaguid(entry, metadata) {
-    const candidates = [entry?.aaguid, metadata?.aaguid];
-    for (const candidate of candidates) {
-        const formatted = formatGuidCandidate(candidate);
-        if (formatted) {
-            return formatted;
-        }
-    }
-    return '';
-}
-
-function formatGuidCandidate(value) {
-    if (value === undefined || value === null) {
-        return '';
-    }
-
-    if (typeof value === 'string') {
-        const trimmed = value.trim();
-        if (!trimmed) {
-            return '';
-        }
-        if (/^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i.test(trimmed)) {
-            return trimmed.toLowerCase();
-        }
-        const clean = trimmed.replace(/[^0-9a-fA-F]/g, '').toLowerCase();
-        if (clean.length === 32) {
-            return `${clean.slice(0, 8)}-${clean.slice(8, 12)}-${clean.slice(12, 16)}-${clean.slice(16, 20)}-${clean.slice(20)}`;
-        }
-        return '';
-    }
-
-    const bytes = extractByteArray(value);
-    if (bytes && bytes.length === 16) {
-        const hex = bytes.map(byte => byte.toString(16).padStart(2, '0')).join('');
-        return `${hex.slice(0, 8)}-${hex.slice(8, 12)}-${hex.slice(12, 16)}-${hex.slice(16, 20)}-${hex.slice(20)}`;
-    }
-
-    try {
-        if (typeof value.toString === 'function') {
-            return formatGuidCandidate(value.toString());
-        }
-    } catch (error) {
-        // Ignore conversion errors.
-    }
-    return '';
-}
-
-function extractByteArray(value) {
-    if (!value) {
-        return null;
-    }
-    if (Array.isArray(value)) {
-        return value.every(item => Number.isInteger(item)) ? value : null;
-    }
-    if (value instanceof Uint8Array) {
-        return Array.from(value);
-    }
-    if (ArrayBuffer.isView(value)) {
-        return Array.from(new Uint8Array(value.buffer, value.byteOffset, value.byteLength));
-    }
-    if (value instanceof ArrayBuffer) {
-        return Array.from(new Uint8Array(value));
-    }
-    return null;
-}
-
-function extractUserVerification(details) {
-    const values = new Set();
-    if (Array.isArray(details)) {
-        details.forEach(group => {
-            if (Array.isArray(group)) {
-                group.forEach(entry => {
-                    if (entry && entry.userVerificationMethod) {
-                        values.add(formatEnum(entry.userVerificationMethod));
-                    }
-                });
-            }
-        });
-    }
-    return Array.from(values).sort((a, b) => a.localeCompare(b));
-}
-
-function extractTransports(metadata) {
-    const infoTransports = extractList(metadata?.authenticatorGetInfo?.transports);
-    const metadataTransports = extractList(metadata?.transports);
-    const combined = new Set([
-        ...infoTransports.map(formatEnum),
-        ...metadataTransports.map(formatEnum),
-    ]);
-    return Array.from(combined).sort((a, b) => a.localeCompare(b));
-}
-
-function extractList(value) {
-    if (!value) {
-        return [];
-    }
-    if (Array.isArray(value)) {
-        return value.filter(Boolean);
-    }
-    return [value];
-}
-
-function normaliseAaguid(value) {
-    const formatted = formatGuidCandidate(value);
-    return formatted ? formatted.toLowerCase() : '';
-}
-
-function formatListValues(value) {
-    return extractList(value)
-        .map(item => formatEnum(item))
-        .filter(text => typeof text === 'string' && text.trim() !== '');
-}
-
-function formatUpv(upv) {
-    const list = Array.isArray(upv) ? upv : upv ? [upv] : [];
-    const formatted = [];
-    list.forEach(item => {
-        if (item && typeof item === 'object') {
-            const major = item.major ?? item.Major;
-            const minor = item.minor ?? item.Minor;
-            if (major !== undefined && minor !== undefined) {
-                formatted.push(`${major}.${minor}`);
-            }
-        }
-    });
-    return formatted;
-}
-
-function formatAuthenticatorAlgorithms(value) {
-    return extractList(value)
-        .map(item => {
-            if (item && typeof item === 'object') {
-                const type = item.type ? String(item.type) : '';
-                const algValue = item.alg !== undefined && item.alg !== null ? item.alg : item.algorithm;
-                const alg = algValue !== undefined && algValue !== null ? String(algValue) : '';
-                if (type && alg) {
-                    return `${type} (${alg})`;
-                }
-                return type || alg;
-            }
-            if (item !== undefined && item !== null) {
-                return String(item);
-            }
-            return '';
-        })
-        .filter(text => typeof text === 'string' && text.trim() !== '');
-}
-
-function formatDetailValue(value) {
-    if (typeof value === 'boolean') {
-        return value ? 'true' : 'false';
-    }
-    if (value === undefined || value === null) {
-        return '—';
-    }
-    if (Array.isArray(value)) {
-        return value.map(item => formatDetailValue(item)).join(', ');
-    }
-    return String(value);
-}
-
-function formatProtocol(protocol) {
-    if (!protocol) {
-        return '';
-    }
-    const normalised = formatEnum(protocol);
-    const compact = normalised.replace(/\s+/g, '');
-    if (/^fido\d$/i.test(compact)) {
-        return compact.toUpperCase();
-    }
-    return normalised;
-}
-
-function normaliseEnumKey(value) {
-    if (value === undefined || value === null) {
-        return '';
-    }
-    return String(value)
-        .trim()
-        .toUpperCase()
-        .replace(/[^A-Z0-9]+/g, '_')
-        .replace(/^_+|_+$/g, '');
-}
-
-function formatEnum(value) {
-    if (!value && value !== 0) {
-        return '';
-    }
-    return String(value)
-        .split(/[_-]/)
-        .map(part => part.trim())
-        .filter(Boolean)
-        .map(part => {
-            if (/^[A-Z0-9]+$/.test(part)) {
-                if (part.length <= 4) {
-                    return part;
-                }
-                const lower = part.toLowerCase();
-                return lower.charAt(0).toUpperCase() + lower.slice(1);
-            }
-            if (/^.*\d.*$/.test(part)) {
-                return part.toUpperCase();
-            }
-            const lower = part.toLowerCase();
-            return lower.charAt(0).toUpperCase() + lower.slice(1);
-        })
-        .join(' ');
-}
-
-function formatCertification(statusReports) {
-    if (!Array.isArray(statusReports) || !statusReports.length) {
-        return { display: '', status: '' };
-    }
-
-    const sorted = [...statusReports].sort((a, b) => {
-        const dateA = Date.parse(a.effectiveDate || '') || 0;
-        const dateB = Date.parse(b.effectiveDate || '') || 0;
-        return dateB - dateA;
-    });
-
-    const latest = sorted[0];
-    if (!latest) {
-        return { display: '', status: '' };
-    }
-
-    const statusRaw = typeof latest.status === 'string' ? latest.status.trim() : '';
-    const statusValue = statusRaw ? statusRaw.toUpperCase() : '';
-    const descriptor = typeof latest.certificationDescriptor === 'string' ? latest.certificationDescriptor.trim() : '';
-    const certificateNumber = typeof latest.certificateNumber === 'string' ? latest.certificateNumber.trim() : '';
-
-    const parts = [];
-    const statusDisplay = statusValue ? formatEnum(statusValue) : '';
-    if (statusDisplay) {
-        parts.push(statusDisplay);
-    }
-    if (descriptor) {
-        parts.push(descriptor);
-    }
-    if (certificateNumber) {
-        parts.push(`(${certificateNumber})`);
-    }
-
-    return {
-        display: parts.filter(Boolean).join(' • '),
-        status: statusValue,
-    };
-}
-
-function latestEffectiveDate(statusReports) {
-    if (!Array.isArray(statusReports) || !statusReports.length) {
-        return '';
-    }
-    const sorted = [...statusReports].sort((a, b) => {
-        const dateA = Date.parse(a.effectiveDate || '') || 0;
-        const dateB = Date.parse(b.effectiveDate || '') || 0;
-        return dateB - dateA;
-    });
-    return sorted[0]?.effectiveDate || '';
-}
-
-function parseIsoDate(value) {
-    if (typeof value !== 'string' || !value.trim()) {
-        return null;
-    }
-    const parsed = new Date(value);
-    if (Number.isNaN(parsed.getTime())) {
-        return null;
-    }
-    return parsed;
-}
-
-function formatDate(value) {
-    if (!value) {
-        return '';
-    }
-    const date = new Date(value);
-    if (Number.isNaN(date.getTime())) {
-        return value;
-    }
-    return new Intl.DateTimeFormat(undefined, {
-        year: 'numeric',
-        month: 'short',
-        day: 'numeric',
-    }).format(date);
-}
-
 if (typeof window !== 'undefined') {
     window.focusMdsAuthenticator = focusAuthenticatorByAaguid;
-}
-
-function decodeBase64Url(value) {
-    let base64 = value.replace(/-/g, '+').replace(/_/g, '/');
-    const padding = base64.length % 4;
-    if (padding) {
-        base64 += '='.repeat(4 - padding);
-    }
-    const binary = atob(base64);
-    const bytes = new Uint8Array(binary.length);
-    for (let i = 0; i < binary.length; i += 1) {
-        bytes[i] = binary.charCodeAt(i);
-    }
-    return new TextDecoder().decode(bytes);
 }
